#9 composableがcomposableではなかったのでserviceとしてまとめた。
